+++ pkg/chains/conversion.go
@@ -1,10 +1,8 @@
 package chains
 
 import (
-	"encoding/hex"
 	"fmt"
 
-	"cosmossdk.io/errors"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 )
@@ -27,27 +25,4 @@ func StringToHash(chainID int64, hash string, additionalChains []Chain) ([]byte,
 		return hash.CloneBytes(), nil
 	}
 	return nil, fmt.Errorf("cannot convert hash to bytes for chain %d", chainID)
-}
-
-// ParseAddressAndData parses the message string into an address and data
-// message is hex encoded byte array
-// [ contractAddress calldata ]
-// [ 20B, variable]
-func ParseAddressAndData(message string) (ethcommon.Address, []byte, error) {
-	if len(message) == 0 {
-		return ethcommon.Address{}, nil, nil
-	}
-
-	data, err := hex.DecodeString(message)
-	if err != nil {
-		return ethcommon.Address{}, nil, errors.Wrap(err, "message should be a hex encoded string")
-	}
-
-	if len(data) < 20 {
-		return ethcommon.Address{}, data, nil
-	}
-
-	address := ethcommon.BytesToAddress(data[:20])
-	data = data[20:]
-	return address, data, nil
 }